\name{isoSIM-package}
\alias{isoSIM-package}
\alias{isoSIM}
\docType{package}
\title{
 This package simulates the generation of an isofemale line
}
\description{

   Version 0.25 - squashed plotting bug
   Version 0.24 - removed Output.cpp
   Version 0.23 - Removed number_of_founders from calc_allele_spectrum
   Version 0.22 - Added save and load functions
   Version 0.21 - changed random-seed management
   Version 0.20 - Removed superfluous code
   Version 0.19 - Removed number_of_founders from Fst and LD code
   Version 0.18 - Start of tracking changes
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Your Name, email optional.

Maintainer: Your Name <your@email.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.
  }
}
