\name{isoSIM-internal}
\title{Internal isoSIM functions}
\alias{create_two_populations_migration_cpp}
\alias{create_two_populations_cpp}
\alias{select_population_cpp}
\alias{test_fish_functions}
\alias{create_iso_female_line}
\alias{findtype}
\alias{hierfstat_basic_stats}
\alias{calculate_summaryStats}
\alias{create_loci_matrix}
\alias{create_pop_class}
\alias{calculate_average_LD}
\alias{calculate_allele_spectrum_cpp}
\alias{calc_heterozygosity_cpp}
\alias{calc_allele_frequencies}
\alias{calc_heterozygosity_indiv}
\alias{calculate_dist_junctions}
\alias{calculate_heterozygosity}
\alias{calculate_heterozygosity_and_freq_table}
\alias{plot_dist_junctions}
\alias{simulate_from_population}
\alias{create_isofemale_line_cpp}
\alias{plot.individual}
\alias{print.population}
\alias{print.individual}
\alias{create_population_cpp}

\description{Internal Guilds functions}
\details{These are not to be called by the user}
\keyword{internal}