\name{isoSIM-package}
\alias{isoSIM-package}
\alias{isoSIM}
\docType{package}
\title{
 This package simulates the generation of an isofemale line
}
\description{
   Version 0.32 - Moved Fish.h code to Fish.cpp \cr
   Version 0.31 - Changed random number generator to R based \cr
   Version 0.30 - Added Recombination = 1 code \cr
   Version 0.29 - Changed internal junction representation: removed .left \cr
   Version 0.28 - Reverted to Agner Fog Random number generation \cr
   Version 0.27 - Speed up return types \cr
   Version 0.26 - Added class verification code \cr
   Version 0.25 - Squashed plotting bug \cr
   Version 0.24 - Removed Output.cpp \cr
   Version 0.23 - Removed number_of_founders from calc_allele_spectrum \cr
   Version 0.22 - Added save and load functions \cr
   Version 0.21 - Changed random-seed management \cr
   Version 0.20 - Removed superfluous code \cr
   Version 0.19 - Removed number_of_founders from Fst and LD code \cr
   Version 0.18 - Start of tracking changes \cr
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Thijs Janzen (thijsjanzen@gmail.com)
}
